refact: Move Processing components into dedicated namespace
